requirements.txt
selenium
requests
pytz
python-dotenv
supabase